major: draft arquitecture creation
It's composed of two systems, game and console.

-- Game
The game has been conceived to be used as a library, which the game DEV will import and
extend, this library allows the user to create a game with the following
features:
    - Sprites
    - Events
        Mouse
        Keyboards
        Collitions
    - Sounds
The general idea is that by combining those features a game can be born,
for example a menu is composed by buttons, and a button could be
definied as a sprite with its collition shape which on mouse hover will
emit a signal.

The library will borrow terminology and certain logic from GODOT, like
the signal system for example.

-- Console
The console will read the game instructions, and execute them. Therefore
it's executing the game. It would be able to understand and interpret
the set of instructions.