chore: interop orientation
Updated system names using a better terminology. The frontend is where
the game will be developed and the backend is a system in charge of
running the game.

- Updated the overall arquitecture allowing both frontend and backend
  coexist by using the frontend function interface of the frontend.

We made several spikes looking to support implicit calls in the target
language, so that the VM won't have to interpret the whole target
language but rather the pieces of code related to make the game run.
The spikes showed that this idea was hard to implement, given the short
time we discarded it.

Then after research we found the concept of Foreign Function Interface,
and the spikes conclusion was successful. That's why I've changed the
Backend to use the FFI instead.

- Added the Player person, which will load a frontend game into the
  console.