refactor: moved game_handler to bridge
The language of the bridge offers a FFI(Foreign Function Interface) that
allows it to call for backend routines. The older arquitecture design had
and issue, it was assuming the backend to offer the FFI for the game to
use; it is an issue because in order to go along with it the bridge has to be refactored
That's why game_loop was moved as a component of bridge.

- Moved the containers output_handler, input_handler and game_handler as
  components of the new console container. And the components of the
  moved containers will pass to Level 4.