refactor: merged console with bridge
console was conceived to handle the low-level logic of the bridge.
This design was meant to fix low-level bugs without having to rebuild
the game. The new design keeps this behaviour if the game is compiled
without the bridge dependency.

Using console as a middle-man between bridge and screen meant an extra
layer of control but the price is the complexity of implementation and
mantenaibility.

It is not cost effective to use a JNI given that there aren't
performance-critical operations to perform by the bridge.

console was removed for the following reasons:
- performance
    marshaling/unmarshaling between console and bridge and then
    repeat the same operations between console and screen.
- complexity
    a independient socket server running alognside the game and screen
    to enable their communication. This means console and socket server
    would need an agreement regarding frames.
- reliability
    On each tick of the game loop, bridge would call console's routines
    to fetch and bring frames to the screen. 60 frames per second at the
    minimum.

The new implemented design
    - removes extra marshaling/unmarshaling operation,
    - reduces the complexity because the socket server and client
      and present inside the bridge;
    - improves relialibilty by allowin the bridge communicate directly
      with screen.